Initial version of Light control plugin. Visualization only.
The light control tool is included as a plugin. To access the tool, click the unlabeled cog button in the toolbar above the level editor.

Tool opens in a dockable tab, with a tree-style view of the lights and an editor for the lights' main properties on the side. There is absolutely no integration between the tool and the actual level shown in the editor, so the tool currently is only a visual prototype.