Hooked up the tree view to the engine, minus removing items from it on actor deletion
Upon creating the window of the tool, it fetches any existing lights in the scene to display them in the tree. In addition, any future lights added into the scene are automatically added to the tree via a callback.

Fixed some errors with the drag-drop operations in the tree which lead to crashes when edge cases were met. If a light is drag dropped on another light, creates a new group and puts both lights in it. The New Group button adds a new group to the tree as well now.

Added checkboxes which control the on/off state of the lights directly in the tree. Groups with no lights in them have no checkboxes, and groups with both active and inactive lights are using an "undetermined" checkbox.

Tree items can be renamed by double-clicking the name. Commit the name with Enter.

Important: While lights are added automatically to the tree, they are not removed automatically because the engine doesn't have a clear way to notify the tool about actor deletion.