Connected more of the engine data to the tool. QoL improvements.
The Hue and Saturation values of a light can be changed via the tool, and it will reflect in the engine. The saturation gradient will also change based on the hue used by the light.

Implemented the ability to save the current state of the light control tool into a .json file, and following that load it back from the same file. Not automated, but the basis is there.  

Started working on quality of life features so that I can design their implementation and apply it to the remaining fields for the lights rather than retroactively having to fix them later.

Researched how the Engine's default icons can be used in the tool, and the results can be seen in the tree.

Will likely take time to split the light control tool into multiple files and classes because the code is beginning to become very unwieldy to work in and maintain.