Modified the tree to include icons from the UE4 editor
Changes were done in attempt to replicate the mock-up. The folder items in the tree use a larger font and use a different layout compared to the lights. The previous checkboxes which controlled the on/off state have been replaced visually by icons representing the type of light. If a folder contains different light types it uses the point light icon due to no better asset at my disposal. If the folder contains both enabled and disabled lights, its icon takes on a yellow-ish tint to signify that.

While the small arrow next to the folders is still available, the folders can now be expanded and closed by pressing on their folder icon on the side.