Improved icons in tree view. Implemented the temperature slider.
Replaced the slate brushes used for the icons in the tree with the icons used in the content browser as they are higher resolution and do not look pixelated when scaled up.

The temperature slider in the property editor has been implemented and works for point lights, directional lights and spotlights. There is no implementation for sky lights because the engine does not support that. Because of this, if a skylight is selected as the master light, the temperature slider and checkbox are disabled.