Finished the light header implementation. Fixed edge cases which caused crashes.
The light header now allows for the light's name and its additional note to be changed via a double click. If there is no additional note associated with the light, uses a placeholder to make it clear one can be added. The additional note disaplays in the tree view as well.

Fixed a bug where a group created when dragging an item on top of a light was not assigned a parent correctly. This would cause the reparenting of this group to crash the engine because it would duplicate it in the tree view and the Slate Tree View class cannot handle that.

The saving and loading now includes all relevant light properties which are implemented in the editor as well.

Including a showcase state (ShowCase.json) which uses most if not all of the features in the tool.