Changed how values are stored in the tree. Implemented orientation controls for spotlights
Added the SLightSpecificProperties class which is responsible for selecting additional UI elements based on the selected master light.  Currently only supports spot lights. Spotlight implementation has sliders for the horizontal and vertical orientation, which maintain relative orientation between the master light and any other lights selected. Also has inner angle and outer angle sliders which dictate the angle of the spotlight. The inner angle can be "locked" to the outer angle, allowing for it to change proportionally to the outer angle.

All values saved in the tree (Intensity, Hue, Saturation, etc.) are not saved as their actual values in their appropriate units, rather than in a float in the [0.0; 1.0] range. Previous .json files will probably not work correctly with this change.