Reworked Save and Load buttons. Added autosaving.
Save and Load buttons now make use of editor icons rather than text. Added a Save As button.

When a preset is saved or loaded, its path is saved and used whenever the user presses the save button or an autosave is triggered. Autosaves currently happen once every 5 minutes because the tool still seems to have a lot of edge cases which cause crashes.

When the tool is closed or an autosave happens, the tool is also going to save the last preset that was used together with its name, and open it the next time the tool opened. If none was selected, resorts to using a default autosave file.