Added Undo/Redo for the light properties.
Renamed FTreeItem to ULightTreeItem and changed the base class from TSharedFromThis<FTreeItem> to UObject. This was done as it is a requirement from the UE4 Transaction system which is used for the undo/redo. This also meant that the items can no longer be passed around with shared ptrs, which caused a lot of code changes.