Fixed oversight which caused engine to crash on level save in Editor. Started work on DMX controls.
Fixed a bug which caused the entire tree for the light control tool to be garbage collected because no items in it were registered to it, resulting in illegal memory deref. Resolved by adding the UObject which holds all of the tree data to the garbage collection root.  

Made first forays into creating a basic DMX control setup. Currently it is only hardcoded to send 128 to the first two channels for test purposes when the output port is selected.

Note: DefaultEngine.ini includes the output port used by the Enttec Ode Mk2. Any other DMX controllers like it would have to be added there with their respective IP address.