Finished first iteration of DMX controls
The DMX controls can now control all variables for the lights. Due to variety in the DMX setups for different lights each light's conversion must be defined manually. This can be done with blueprints, see Confluence.

Included here is a blueprint which does the conversion for the Algam Miniwash 710 light.