Added new asset type for DMX light configs. Added a new window for the DMX tool.
Implemented all the necessary classes to allow for the creation of a custom asset type called DMX Light Config, which contains info about the mapping of light properties to DMX channels. 

Did some initial work the window for the DMX controls, but I will likely modify the virtual lights tool code to make it reusable with the DMX lights.