Reworked the basis of the virtual light control tool fpr better modularity
Moved the data of the tool out of the SLightTreeHierarchy class and into a separate UToolData class. The items in the tree are now of type UItemHandle, with each handle having a UBaseLight*. This pointer it meant to be of the type expected by the tool.

Respectively, the code for all other widgets has been changed to accommodate this change. Code specific to virtual lights which they contained has been moved to SLightControlTool and is provided to the widgets via delegates. This way the widgets can be reused with data for DMX lights without taking all the same steps as with the virtual lights.