The DMX control tool now properly sends signals to DMX fixtures based on the config assets assigned.
Added GetIntensityNormalized, GetTemeperatureNormalized, GetHorizontalNormalized and GetVerticalNormalized to UBaseLight as a way for the different light types to define their own ranges for these variables without the slate widgets using the normalized values needing to account for the type of light that is being used. In the same regard, AddHorizontal and AddVertical now use values between 0 and 1 as opposed to degrees between -180 and 180.

The UDMXConfigAsset was expanded to include all necessary channels which could be easily tested out, as well as an array of channels which can be kept constant.

The DMX Control tool uses the same layout as the virtual light control tool. Main difference is that instead of the transform viewer, it has DMX properties, such as the starting channel for the fixture, what config to use for the channels, and what port to use for the output.