First implementation of a light gel palette for the light tools.
The gel palette is a separate widget which gets its colors from a .txt of values taken from https://www.leefilters.com/lighting/colour-list.html# . Hovering over a color in it displays the HSV of the color.

For now the palette is opened in a separate window, by a button that is part o f the light properties widget. Subject to change.