PhysicalObjectTrackingComponent now saves device by Serial ID
Also fixed a bug where the shake detection task would not detect devices identified as controllers due to the code overwriting their indices.

The Virtual Light Controls and DMX Light Controls are now initialized with the module startup, rather than when the user requests them to be opened. Similarly, closing the window will not delete the data associated with the tools. This was done for future blueprint integration of the tools.