Light control plugin now automatically disables slate throttling for the project
The throttling is a variable that is not available via the project settings window, and is instead only accessible via the editor console. It appears to be enabled by default, in which case the viewport will freeze anytime a UI widget is interacted with using the mouse, e.g. dragging a slider. This makes it difficult to edit the lights via the tool, which is why I opted to disable it automatically in the plugin.